// src/app/modules/case-management/models/viewModels/FormSectionStatesModel.ts
export interface FormSectionStatesModel {
    finalReviewScore: string
    reviewScoreBenchmark: string
    reviewMarkButtonLabel: string
    reviewMarkLoading: boolean
    reviewMarkedReadCounter: number
    counterTooltipVisible: boolean
    reviewMarkedRead: boolean
    reviewStartLoading: boolean
    reviewCancelLoading: boolean
}